feat: Integrate payment records and visibility logic for Pay button in PayableDetails
