+++ packages/sdk-react/src/components/payables/PayableDetails/PayableDetails.tsx
@@ -57,7 +57,7 @@ const PayableDetailsBase = ({
     lineItems,
     isEdit,
     isLoading,
-    isProcessingPayment,
+    showPayButton,
     actions: {
       setEdit,
       createInvoice,
@@ -153,8 +153,8 @@ const PayableDetailsBase = ({
           payInvoice={payInvoice}
           payableDetailsFormId={payableDetailsFormId}
           onClose={onClose}
-          isProcessingPayment={isProcessingPayment}
           modalComponent={modalComponent}
+          showPayButton={showPayButton}
         />
         <DialogContent sx={{ display: 'flex', flexDirection: 'column' }}>
           <Grid container columnSpacing={4} height="100%">
